Alterações para corrigir erro conta corrente

+++ src/main/java/br/edu/ifba/inf012/internetBanking/dtos/contaCorrente/ContaDto.java
@@ -2,7 +2,7 @@ package br.edu.ifba.inf012.internetBanking.dtos.contaCorrente;
 
 import br.edu.ifba.inf012.internetBanking.models.ContaCorrente;
 
-public record ContaDto(Long id, int numero, int agencia, String saldo, String usuario) {
+public record ContaDto(Long id, Long numero, int agencia, String saldo, String usuario) {
 	public ContaDto(ContaCorrente conta) {
 		this(conta.getId(), conta.getNumero(), conta.getAgencia(), conta.getSaldoString(), conta.getUsuario().getNome());
 	}
